This commit successfully add google login with issues on facebook's

+++ project.py
@@ -4,7 +4,7 @@ from flask import Flask, render_template, request, redirect,jsonify, url_for, fl
 
 from sqlalchemy import create_engine, asc
 from sqlalchemy.orm import sessionmaker
-from database_setup import Base, Restaurant, MenuItem
+from database_setup import Base, Restaurant, MenuItem, User
 
 # NEW IMPORTS FOR THIS STEP
 from flask import session as login_session
@@ -33,7 +33,7 @@ APPLICATION_NAME = "Restaurant Menu Application"
 
 
 #Connect to Database and create database session
-engine = create_engine('sqlite:///restaurantmenu.db')
+engine = create_engine('sqlite:///restaurantmenuwithusers.db')
 Base.metadata.bind = engine
 
 DBSession = sessionmaker(bind=engine)
@@ -53,6 +53,102 @@ def showLogin():
     # return "The current session state is %s" %login_session['state']
     # STATE=state was later added after being created in login.html
     return render_template('login.html', STATE=state)
+
+
+# FACEBOOK SIGN IN
+@app.route('/fbconnect', methods=['POST'])
+def fbconnect():
+    if request.args.get('state') != login_session['state']:
+        response = make_response(json.dumps('Invalid state parameter.'), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+    access_token = request.data
+    print "access token received %s " % access_token
+    # Below, exchange client token for long-lived server side token
+    # with GET /oauth/access_token?grant_type=fb_exchange_token&client_id=
+    # {app-id}&client_secret={app-secret}&fb_exchange_token={short-lived-token}
+    app_id = json.loads(open('fb_client_secrets.json', 'r').read())[
+        'web']['app_id']
+    app_secret = json.loads( # I have to send my app secret to Facebook
+        open('fb_client_secrets.json', 'r').read())['web']['app_secret']
+    url = 'https://graph.facebook.com/oauth/access_token?grant_type=fb_exchange_token&client_id=%s&client_secret=%s&fb_exchange_token=%s' % (
+        app_id, app_secret, access_token)
+    h = httplib2.Http()
+    result = h.request(url, 'GET')[1]
+
+
+    # Use token to get user info from API
+    userinfo_url = "https://graph.facebook.com/v2.8/me"
+    '''
+        Due to the formatting for the result from the server token exchange we have to
+        split the token first on commas and select the first index which gives us the key : value
+        for the server access token then we split it on colons to pull out the actual token value
+        and replace the remaining quotes with nothing so that it can be used directly in the graph
+        api calls
+    '''
+    token = result.split(',')[0].split(':')[1].replace('"', '')
+
+    url = 'https://graph.facebook.com/v2.8/me?access_token=%s&fields=name,id,email' % token
+
+    h = httplib2.Http()
+    result = h.request(url, 'GET')[1]
+    # print "url sent for API access:%s"% url
+    # print "API JSON result: %s" % result
+    data = json.loads(result)
+    login_session['provider'] = 'facebook'
+    login_session['username'] = data["name"]
+    login_session['email'] = data["email"]
+    login_session['facebook_id'] = data["id"]
+
+    # The token must be stored in the login_session in order to properly logout
+    login_session['access_token'] = token
+
+    # Get user picture
+    url = 'https://graph.facebook.com/v2.8/me/picture?access_token=%s&redirect=0&height=200&width=200' % token
+    h = httplib2.Http()
+    result = h.request(url, 'GET')[1]
+    data = json.loads(result)
+
+    login_session['picture'] = data["data"]["url"]
+
+    # see if user exists
+    user_id = getUserID(login_session['email'])
+    if not user_id:
+        user_id = createUser(login_session)
+    login_session['user_id'] = user_id
+
+    # Welcome splash screen
+    output = ''
+    output += '<h1>Welcome, '
+    output += login_session['username']
+
+    output += '!</h1>'
+    output += '<img src="'
+    output += login_session['picture']
+    output += ' " style = "width: 300px; height: 300px;border-radius: 150px;-webkit-border-radius: 150px;-moz-border-radius: 150px;"> '
+
+    flash("Now logged in as %s" % login_session['username'])
+    return output
+
+
+@app.route('/fbdisconnect')
+def fbdisconnect():
+    facebook_id = login_session['facebook_id']
+    # The access token must me included to successfully logout
+    access_token = login_session['access_token']
+    url = 'https://graph.facebook.com/%s/permissions?access_token=%s' % (facebook_id,access_token)
+    h = httplib2.Http()
+    result = h.request(url, 'DELETE')[1]
+    del login_session['username']
+    del login_session['email']
+    del login_session['picture']
+    del login_session['user_id']
+    del login_session['facebook_id']
+    return "you have been logged out"
+
+
+
+
 
 
 # HANDLER OF CODE SENT BACK FROM CALLBACK METHOD
@@ -92,7 +188,7 @@ def gconnect():
     h = httplib2.Http()
     result = json.loads(h.request(url, 'GET')[1])
 
-    # If there was an error in the access token info, abort
+    # This code: If there was an error in the access token info, abort
     # if x is not None:
         # # Do something about x
     if result.get('error') is not None:
@@ -101,6 +197,7 @@ def gconnect():
         return response
 
     # Verify that the access token is used for the intended user.
+    # Only the it_token part is extracted from credential object.
     gplus_id = credentials.id_token['sub']
     if result['user_id'] != gplus_id:
         response = make_response(
@@ -140,6 +237,26 @@ def gconnect():
     login_session['picture'] = data['picture']
     login_session['email'] = data['email']
 
+
+    #LOCAL PERMISSION
+    # OBTAINING CREDENTIALS OF AN EXISTING USER.
+    # If a user exists, if it doesn't make a new one.
+    # Lets run getUserID on the email address stored in our
+    # login_session.
+    user_id = getUserID(login_session['email'])
+    # If user_id doesn't have a value
+    if not user_id:
+        # Make a new user account for her locally
+        # and pass in the login_session.
+        user_id = createUser(login_session)
+        # After either making a new account or
+        # retreiving an existing one, we store her user ID
+        # in the login_session under user_id.
+        #The left side is variable. The right is the assignment.
+    login_session['user_id'] = user_id
+    # END OF LOCAL PERMISSION
+
+
     output = ''
     output += '<h1>Welcome, '
     output += login_session['username']
@@ -150,6 +267,57 @@ def gconnect():
     flash("you are now logged in as %s" % login_session['username'])
     print "done!"
     return output
+
+
+# LOCAL PERMISSION SYSTEM
+# User Helper Functions
+
+# createUser takes in login_session as input
+def createUser(login_session):
+
+    """create new user in our database, extracting all
+    the fields neccessary to populate it from information
+    gathered from the login_session"""
+
+    newUser = User(name=login_session['username'],
+        email=login_session['email'], picture=login_session['picture'])
+    session.add(newUser)
+    session.commit()
+    user = session.query(User).filter_by(email=login_session['email']).one()
+    # Then returns a user_id of the new user created
+    return user.id
+
+
+
+def getUserInfo(user_id):
+
+    """If a user ID is passed into this method,
+    it simply returns the user object associated with this ID number."""
+
+    user = session.query(User).filter_by(id=user_id).one()
+    # Returns user object associated with this number.
+    return user
+
+
+
+def getUserID(email):
+
+    """This method, takes an email address and return and ID,
+    if that email address belongs to  user stored in our database"""
+
+    try:
+        user = session.query(User).filter_by(email=email).one()
+        # Returns an ID number if the email address belongs to
+        # a user stored in our database.
+        return user.id
+    except:
+        # If not, it returns None.
+        return None
+
+# END OF LOCAL PERMISSION
+
+
+
 
 
 # DISCONNECT - Revoke a current user's token and reset their login_session
@@ -179,7 +347,7 @@ def gdisconnect():
         response.headers['Content-Type'] = 'application/json'
         return response
     else:
-        response = make_response(json.dumps('Failed to revoke token for given user.', 400))
+        response = make_response(json.dumps('Failed to revoke token for given   user.', 400))
         response.headers['Content-Type'] = 'application/json'
         return response
 
@@ -208,18 +376,24 @@ def restaurantsJSON():
 @app.route('/restaurant/')
 def showRestaurants():
   restaurants = session.query(Restaurant).order_by(asc(Restaurant.name))
-  return render_template('restaurants.html', restaurants = restaurants)
+  if 'username' not in login_session:
+      return render_template('publicrestaurants.html', restaurants = restaurants)
+  else:
+      return render_template('restaurants.html', restaurants = restaurants)
+
 
 #Create a new restaurant
 @app.route('/restaurant/new/', methods=['GET','POST'])
 def newRestaurant():
-    # Protect the page
+    # ADD LOGIN PERMISSION
     # If a user name is not detected for a given request.
     # Lets redirect to login page.
     if 'username' not in login_session:
         return redirect('/login')
     if request.method == 'POST':
-        newRestaurant = Restaurant(name = request.form['name'])
+        newRestaurant = Restaurant(name = request.form['name'],
+            # Create the user_id field when you create a new Restaurant MenuItem
+            user_id=login_session['user_id'])
         session.add(newRestaurant)
         flash('New Restaurant %s Successfully Created' % newRestaurant.name)
         session.commit()
@@ -230,9 +404,13 @@ def newRestaurant():
 #Edit a restaurant
 @app.route('/restaurant/<int:restaurant_id>/edit/', methods = ['GET', 'POST'])
 def editRestaurant(restaurant_id):
-  editedRestaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
+  editedRestaurant = session.query(Restaurant).filter_by(id=restaurant_id).one()
   if 'username' not in login_session:
-      return redirect('/login')
+     return redirect('/login')
+     # ADD ALERT MESSAGE TO PROTECT
+  if editedRestaurant.user.id != login_session['user_id']:
+     return "<script>function myFunction() {alert('You are not authorized to edit this restaurant. Please create your own restaurant in order to edit.');}</script><body onload='myFunction()'>"
+
   if request.method == 'POST':
       if request.form['name']:
         editedRestaurant.name = request.form['name']
@@ -248,6 +426,10 @@ def deleteRestaurant(restaurant_id):
   restaurantToDelete = session.query(Restaurant).filter_by(id = restaurant_id).one()
   if 'username' not in login_session:
       return redirect('/login')
+      # ADD ALERT MESSAGE TO PROTECT
+  if restaurantToDelete.user_id != login_session['user_id']:
+      return "<script>function myFunction() {alert('You are not authorized to delete this restaurant. Please create your own restaurant in order to delete.');}</script><body onload='myFunction()'>"
+
   if request.method == 'POST':
     session.delete(restaurantToDelete)
     flash('%s Successfully Deleted' % restaurantToDelete.name)
@@ -261,8 +443,16 @@ def deleteRestaurant(restaurant_id):
 @app.route('/restaurant/<int:restaurant_id>/menu/')
 def showMenu(restaurant_id):
     restaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
+    # To protect each menu based on whoever created it.
+    creator = getUserInfo(restaurant.user_id)
     items = session.query(MenuItem).filter_by(restaurant_id = restaurant_id).all()
-    return render_template('menu.html', items = items, restaurant = restaurant)
+    # If a user isn't logged in or isn't the original creator
+    if 'username' not in login_session or creator.id !=login_session['user_id']:
+        return render_template('publicmenu.html',
+            items=items, restaurant=restaurant, creator=creator)
+    else:
+        # Otherwise they are logged in and are the creator.
+        return render_template('menu.html', items = items, restaurant = restaurant, creator=creator)
 
 
 
@@ -272,8 +462,13 @@ def newMenuItem(restaurant_id):
   restaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
   if 'username' not in login_session:
       return redirect('/login')
+      # ADD ALERT MESSAGE TO PROTECT
+  if login_session['user_id'] != restaurant.user_id:
+      return "<script>function myFunction() {alert('You are not authorized to add menu items to this restaurant. Please create your own restaurant in order to add items.');}</script><body onload='myFunction()'>"
+
   if request.method == 'POST':
-      newItem = MenuItem(name = request.form['name'], description = request.form['description'], price = request.form['price'], course = request.form['course'], restaurant_id = restaurant_id)
+      newItem = MenuItem(name = request.form['name'], description = request.form['description'], price = request.form['price'], course = request.form['course'],
+            restaurant_id = restaurant_id, user_id=restaurant.user_id)
       session.add(newItem)
       session.commit()
       flash('New Menu %s Item Successfully Created' % (newItem.name))
@@ -289,6 +484,10 @@ def editMenuItem(restaurant_id, menu_id):
     restaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
     if 'username' not in login_session:
         return redirect('/login')
+        # ADD ALERT MESSAGE TO PROTECT.
+    if login_session['user_id'] != restaurant.user_id:
+        return "<script>function myFunction() {alert('You are not authorized to edit menu items to this restaurant. Please create your own restaurant in order to edit items.');}</script><body onload='myFunction()'>"
+
     if request.method == 'POST':
         if request.form['name']:
             editedItem.name = request.form['name']
@@ -313,6 +512,10 @@ def deleteMenuItem(restaurant_id,menu_id):
     itemToDelete = session.query(MenuItem).filter_by(id = menu_id).one()
     if 'username' not in login_session:
         return redirect('/login')
+        # ADD ALERT MESSAGE TO PROTECT
+    if login_session['user_id'] != restaurant.user_id:
+        return "<script>function myFunction() {alert('You are not authorized to delete menu items to this restaurant. Please create your own restaurant in order to delete items.');}</script><body onload='myFunction()'>"
+
     if request.method == 'POST':
         session.delete(itemToDelete)
         session.commit()
@@ -321,6 +524,28 @@ def deleteMenuItem(restaurant_id,menu_id):
     else:
         return render_template('deleteMenuItem.html', item = itemToDelete)
 
+
+# Disconnect based on provider
+@app.route('/disconnect')
+def disconnect():
+    if 'provider' in login_session:
+        if login_session['provider'] == 'google':
+            gdisconnect()
+            del login_session['gplus_id']
+            del login_session['access_token']
+        if login_session['provider'] == 'facebook':
+            fbdisconnect()
+            del login_session['facebook_id']
+        del login_session['username']
+        del login_session['email']
+        del login_session['picture']
+        del login_session['user_id']
+        del login_session['provider']
+        flash("You have successfully been logged out.")
+        return redirect(url_for('showRestaurants'))
+    else:
+        flash("You were not logged in")
+        return redirect(url_for('showRestaurants'))
